Refactor test setup
