+++ jest.config.js
@@ -79,7 +79,8 @@ module.exports = {
     // notifyMode: "always",
 
     // A preset that is used as a base for Jest's configuration
-    // preset: null,
+
+    preset: 'jest-puppeteer',
 
     // Run tests from one or more projects
     // projects: null,
