+++ package.json
@@ -41,9 +41,10 @@
     "build": "react-scripts build",
     "test:local": "react-scripts test",
     "test:coverage": "npm run test:unit --coverage",
-    "test:unit": "cross-env CI=true react-scripts test --env=jsdom",
-    "test": "npm run test:unit && npm run test:e2e",
-    "test:e2e": "rimraf ./user-data && jest -c ./test/e2e_tests/login.jest.config.js && jest -c ./test/e2e_tests/jest.config.js",
+    "test:unit": "cross-env CI=true npm test --env=jsdom -- --testPathIgnorePatterns=e2e",
+    "test": "react-scripts test",
+    "test:all": "npm run test:unit --coverage && npm run test:e2e",
+    "test:e2e": "rimraf ./user-data && npm test -- --testPathPattern=e2e --runInBand",
     "test:watch": "react-scripts test --watchAll",
     "eject": "react-scripts eject",
     "lint": "eslint src --fix",
