+++ src/functional_components/ContactList/contactList.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import TestRenderer from 'react-test-renderer';
 import ContactList from './ContactsList';
-import { contacts } from '../../../test/fixtures/contacts';
+import { contacts } from '../../test/fixtures/contacts';
 
 describe('ContactList', () => {
     test('should render correctly', () => {
