+++ src/test/e2e_tests/config.js
@@ -1,0 +1,13 @@
+require('dotenv').config();
+
+export const launchConfig = {
+    headless: false,
+    slowMo: 0,
+    defaultViewport: {
+        width: 1920,
+        height: 1024,
+    },
+    userDataDir: './user-data',
+};
+
+export const testServerUrl = 'http://localhost:3000';
