+++ src/test/e2e_tests/e2e.test.js
@@ -1,0 +1,3 @@
+require('./login');
+require('./drive');
+require('./editProfile');
