+++ src/test/e2e_tests/login.js
@@ -1,9 +1,14 @@
 require('dotenv').config();
+const puppeteer = require('puppeteer');
+import { launchConfig, testServerUrl } from './config';
 
 jest.setTimeout(16000);
+console.log(launchConfig);
 describe('login', () => {
     test('should go through login process', async () => {
-        await page.goto('http://localhost:3001');
+        const browser = await puppeteer.launch(launchConfig);
+        const page = await browser.newPage();
+        await page.goto(testServerUrl);
         await page.waitForSelector('[data-test-id=login_btn]');
         await page.click('[data-test-id=login_btn]');
         await page.waitForSelector('[data-test-id="solid.community"');
@@ -20,5 +25,6 @@ describe('login', () => {
             (e) => e.innerHTML
         );
         expect(header).toBe('Drive');
+        browser.close();
     });
 });
