--- test/e2e_tests/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-    preset: 'jest-puppeteer',
-    testRegex: './*\\.test\\.js$',
-    testPathIgnorePatterns: ['./login.test.js'],
-};
