--- test/e2e_tests/login.jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    preset: 'jest-puppeteer',
-    testRegex: 'login\\.test\\.js',
-};
